// src/components/data/async_data/async_iterable_resolver.d.ts
import { default as React } from 'react';
/**
 * AsyncIterableResolverProps represents the props that can be passed to the
 * AsyncIterableResolver.
 */
export interface AsyncIterableResolverProps<T, R = any, N = any> {
    asyncIterable: AsyncIterable<T, R, N>;
    next?: N;
    children?: React.ReactNode | React.ReactNode[];
}
/**
 * AsyncIterableResolver is a component that resolves an AsyncIterable into a series of AsyncSnapshots.
 * This component is useful for consuming an AsyncIterable in a React component.
 */
declare const AsyncIterableResolver: React.FC<AsyncIterableResolverProps<unknown>>;
export default AsyncIterableResolver;